requirements.txt
numpy
robin-stocks
python-dotenv
pandas
openpyxl
matplotlib
xlrd